io: remove dead branch from read-only TMemFile constructor
ParseOption("READ") returns EMode::kRead, so we know that
the if will never execute, and that we do not need to create
the file, since we are reusing external data. Nevertheless,
make the file a zombie if the input buffer is empty, since
a read-only empty file is otherwise useless.