[TMVA] Fix GUI can't find all trained BDT's for plots (#3253)
* TMVA-GUI: fix group folder name for methods

When writing trained methods to the output file, there was a typo in the
target directory. This lead to the results of a  method of type "BDT" and
name "MyBDT" being placed (incorrectly) in a folder named
"Method_MyBDT/MyBDT".

This patch places the data correctly in "Method_BDT/BDT". This is
important since e.g. the GUI relies on all BDT's being placed in the
"MethodBDT" folder.

Squashed commits (slight refactoring and fixups):
* TMVA: MethodBase -- fix indentation isses
* TMVA: MethodBase -- refactor datasetName
* TMVA: fix line wrapping
* TMVA: refactor for clarity and better naming
* clang-format
* fix typo in type
* `fFactoryBaseDir` is _not_ a field of MethodBase
* Fix directory name in envelope test