fitting_histograms.md formatting (#3541)
* Fix typo with code tags in FittingHistograms.md

* Standardise presentation of TH1 and TF1 in FittingHistograms.md

There were multiple different presentations of TH1, TF1 and functions e.g.
TH1::Fit. Some TH1 and TF1 references were in bold, others were not. Likewise,
sometimes the TH1 in e.g. TH1::Fit was in bold while sometimes it was not.

This commit changes all instances of TH1, TF1 which are alone to bold and in
code tags. All instances of TH1:: or TF1:: are not in bold.