Fix interpretation of RooAddPdf coefficients for convolutions.
When using RooAddPdfs in a convolution, it has to be ensured that the convolution observable remains the reference for interpreting the coefficients of the RooAddPdf. This is because a new binned range has to be constructed to perform discrete FFTs. Interpreting coefficients with respect to this internal range yields coefficients that depend on the properties of this range. This patch disentangles these.

ROOT-9653
ROOT-7183
ROOT-9419