[DF] Defer declaration of type aliases for Define'd values
Instead of declaring types of Define'd values to the interpreter
on the spot, we now wait until just before the event loop, and jit
these declarations together with everything else.
Not only this is faster (few calls to the interpreter with more code
to jit rather than many calls with little code), but allows RDF to
not call the interpreter at all if no node in the computation graph
is just-in-time compiled, which resolves ROOT-9707.