[cxxmodules] Do not attach an overlay file if no -cxxmodule is specified.
We cannot merge both codepaths because root.exe should always run with modules
if -Druntime_cxxmodules is specified, however, rootcling enables modules only
if -cxxmodule flag is passed.

This patch fixes asserts when ROOT is built in -Druntime_cxxmodules=On but
no -cxxmodule flag is specified (in tree/test for instance).