[tcling] If we fail to dlopen at that stage is a hard error.
In the past we used dlopen to check if the library candidate has shared object
format. Now this check is moved before and if we fail to dlopen we should
really complain. We intentionally dlopen the same library to aid our JIT
symbol resolution during static initialization. We want to diagnose if
the process was not successful at that stage because later it will yield
unresolved JIT symbols which are hard to debug.