[cxxmodules] Decouple loading of modulemaps and adding prebuilt module paths.
When calling clang::HeaderSearch::lookupModule clang can enable
`AllowExtraModuleMapSearch` which looks into the subdirectories to look for
modulemaps. This can leave us in a situation where we have loaded the file
outside of the context of `RegisterPrebuiltModulePath` and when the routine is
called does not add the relevant prebuilt module path forcing clang to build
the module instead of just loading it.

This patch allows the prebuilt module path to still be registered even though
the modulemap file was loaded at an earlier stage.