[RF] Fix writing of large RooTreeDataStores.
RooFit has the option to hold fitting data in memory as std::vectors or
to use Tree-backed storage. The latter only makes sense when streaming
from files is desired, and is therefore not the default.
However, Tree-backed storage was using memory-resident trees, which
could not be written when larger than 1 Gb. Now, the tree-backed storage
is file-backed, and stays connected to the file when reading back.

This fixes ROOT-10188.