Fix root-config now that 'thread' is always enabled
The 'thread' option could not be disabled, so it has been removed
by commit 8e0e72cfd2e8aaa3b446cf997d21db61c3323879. However, its
value was being used in root-config in a way that was not caught
by the initial inspection. The effect was to drop the -pthread flag
from root-config --cflags, which is now fixed by this commit.