[PyROOT] Class `ndarray` to global scope
This PR allows for distributing `AsNumpy` operations to a Spark cluster.

In the implementation of `AsNumpy` function in the pythonizations of RDataFrame,
the class `ndarray` (derived from numpy.ndarray) is created in the scope of the function.

Unfortunately this clashes with a possible implementation of `AsNumpy` in a distributed environment, since pickling of classes inside function scopes is not possible (https://docs.python.org/3/library/pickle.html#pickle-picklable).

To make life easier, the class `ndarray` has been moved to the global scope of the python module.
No functionality has been changed or added, the only further requirement is that now `numpy` has to be imported at the module level.