[PyROOT][ROOT-10268] Enable pickling of arrays returned by AsNumpy
We renamed the class of the numpy wrapper object to numpy.array so that
the wrapping is hidden from the user. However, this breaks pickling of
the object since the class name does not match the name of the true
class. Removing the renaming solves the issue. Now, the returned array
is not called numpy.array but ndarray.