[tcling] Skip weak undefined symbols produced by gcc.
We get these symbols even for a simple hello world message. While it is not
clear if we can skip all weak undefined symbols, we can certainly skip those
which get resolved to libgcj.so and libitm.so.

Oddly enough gcc emits a weak undefined symbol to _Jv_RegisterClasses
(resolved in libgcj.so) which is some gcc/java library.
_ITM_deregisterTMCloneTable and _ITM_registerTMCloneTable are emitted
because (resolved in libitm.so) of pointer arithmetics for transactional
memory support.

The current understanding is that we can safely omit these when harvesting
library dependencies.

This should fix the rootbench builds.