Use dictionary object file as dependency for the PCH
Adding the dictionary target as a dependency for the PCH will not
cause the PCH to regenerate even when the dictionary itself is
regenerated and recompiled. This is due to the fact that CMake does
not create a file-level dependency between the dictionary source and
the PCH. However, adding the dictionary source as dependency creates
a race condition on the custom command that generates the dictionary
source, since both the library owning the dictionary and the PCH can
trigger the custom command independently. Now that we moved to use
object libraries for the dictionaries to workaround another bug in
CMake (namely, https://cmake.org/Bug/view.php?id=14633), we can use
the output object file to create a file-level dependency between the
PCH and the dictionary without introducing a race condition.