Fix tensor allocations in the optimiser classes. Use  tensor allocated in the constructors. Fix CudaBuffer function to copy input from a tensor input Fix formatting (do not use Real_t, i.e. float as template parameter)
restore previous version. New version requires changes in DataLoader to support arrays

Fix test utility function to randomize a tensor. This was providing zero input to some tests and cause them to fails. Add a constructor as in TCudaTensor also in TCPuTensor from just a double/float pointer and shape