[modules] Prefer the headers from $ROOTSYS/include.
If we have -Icomponent/inc and -I$ROOTSYS/include we first try to resolve the
header A from component/inc. However if header B from $ROOTSYS/include
activated a module containing header A we get a redefinition errors.

Those errors are in a way correct because we cannot possibly know that two
headers with the same name coming from two different locations are the same.

This patch give a preference to $ROOTSYS/include and once we move away from
copying headers to a common directory we will rework that part of the modules
support when -Dcxxmodules=On is specified.

This should fix the clangHEAD nightly builds.