[rootcling] Support umbrella headers for dictionaries.
When calling rootcling we expect a set of header files and a selection file.
In some cases it is more convenient to include all of them in a single header
and give it to rootcling. In cases where the header files are so many and
exceed the maximum characters allowed by the console the umbrella header is
the only way to work it around.

In cases where the umbrella header is temporary (such as in cmssw) we cannot
add it as a part of the module, however, we have added its direct includes
in the module definition. Then rootcling should not complain if the umbrella
header file is not part of the module definition.

This patch should fix the issues in cmssw when building a dictionary for
DataFormats/StdDictionaries.