Dot help v2 (#4529)
* Modify ".help" command to open online reference guide:

creates links for class, method and namespace with hash anchors appropriate for doxygen.

Fixing the previous commit - instead of checking if a method is
from the class and then check if it's from its base classes, we directly
check from which class the method is.

Adding the ".help NameSpade::Function", ".help ClassName::Enumeration",
".help ClassName::Enumerator".

Removing all debug text. Adding the handle for functions with enumerators
as a return type or argument. Adding not full solution for the anonymous
enumeration.

Corrected code with respect to most of the comments (exept 3, which will be
in the next commit). There is a lot of modifications - the variables in the
functions are references, the code has better documentation and the variables
are renamed. Added two enumerations and combined 4 functions into 2.

Corrected the code with respect to most of the new comments. Reversed the
logic in "Help" checking. Combined class and namespace finctions with
the help of enumerator "EUrl". Fixed the documentation.

* Test vararg signature building.

With the new features for consteptx and corrected
arguments (",..." - missing).
Still need to be done: Templates, global functions,
data members.

* [base] Refactor .help, add support for structs and data members.

* Update core/base/src/TApplication.cxx

* Fixing compilation and github review comments.

* Fixing compilation and github review comments.

* cleanup

* Fix MacOs flag name

* Fix several typos

* Fix some formatting issues

* More formatting

* Try to fix the Windows error.

Formatting.

* Formatting

* Formatting and use "raw string literal" for re4