[Exp PyROOT] Eggs created in a less pythonic way if requisites not satisfied
Python eggs for PyROOT and Cppyy cannot be created in a complete and
pythonic way if setuptools is not installed and version is > 3.3. The
absence of this conditions actually also make the entire build fail.
If the requisites are not satisfied, eggs are created in a simpler way (with less information).