[Exp PyROOT] Syntax change due to backward incompatibility with Cppyy-1.5.6
As stated in

https://bitbucket.org/wlav/cppyy/issues/176/issue-with-python-list-created-from-vector

in Cppyy-1.5.6 when we do the following:

l = [e for e in cppyy.gbl.get_vec()]

where get_vec() is a C++ injected function that returns a C++
std::vector<std::string>, the variable 'e' that goes through the vector is
not a Python string, but a temporary object of type std::string.

Being 'e' a reference to a temporary, the vector goes out of scope.

The second of the two solutions suggested in the discussion is applied
where necessary.