[PyROOT] Issue a deprecation warning for ROOT.Long and ROOT.Double
As discussed in ROOT-10418, ROOT.Long and ROOT.Double cannot be
used anymore in new Cppyy to pass-by-ref ints and floats,
respectively. Now, the corresponding ctypes types must be used.

This commit sets a deprecation warning in old PyROOT when
trying to attempt such conversion, in preparation for the
switch to the new PyROOT, which will not allow it.