[Exp PyROOT] New exception handler - cppyy patch
Based on:
https://bitbucket.org/wlav/cppyy-backend/src/master/cling/patches/signaltrycatch.diff

Adds a new exception handler used by cppyy to provide a better
debugging experience in presence of errors:
https://cppyy.readthedocs.io/en/latest/debugging.html

It also adds SIGABRT and provides support for signal handling
on Windows.

The original cppyy patch has been modified to make it backwards
compatible, since it eliminated the use of gApplication for
exception handling in Unix. In this version, if there is an
exception handler (e.g. we are using PyROOT, which defines it
via cppyy) we will rely on that handler to treat the exception,
otherwise we fall back to the previous behaviour which relied on
gApplication.