[ROOT-10474] GCC9: Lazily add comparison operators for iterators
In GCC9, operator== and operator!= are no longer member functions
of STL iterators (seen in _Rb_tree_iterator). This means they
cannot be obtained anymore with GetListOfMethods of the iterator
class.

Such change broke the iteration of STL classes from Python when
using (old) cppyy alone in GCC9. The addition of the operators to
the iterator class still happened when importing ROOT because
in that case gApplication is initialized to TPyROOTApplication,
which is required in Utility::AddBinaryOperator to add the
operators.

This commit adds the necessary logic so that operator== and
operator!= are also added to the iterator proxy class when
using cppyy alone in GCC9. The addition is done lazily as a last
attempt before throwing an error in the comparison.