[v7] add ROOT base exception and RStatus class
The RException class is supposed to serve as base class for all ROOT
exceptions.  The RStatus class can be used by infrastructure code that
is supposed to be deployed both in an exception context and in a return
code context.  RStatus is a templated class that captures a return value
(e.g. a bool or an int from a system call).  The captured value can
indicate an error.  In this case, RStatus normally throws an exception.
That behavior can be changed such that the error code is returned
instead.  If the error code remains mistakenly unchecked, an exception
is thrown nevertheless when the RStatus class goes out of scope.