Make bin border comparison specializable, drop operator==(Eq, Eq)
The proposal to drop operator== comes from the fact that this operator
currently does not compare everything (binning, but not title), which is
a can of worm. Operator overloading should only be used in simple cases
where the operator has an unsurprising and unambiguous meaning, which in
this case would be "compare every axis property".