[Exp PyROOT] Exploit lifeline when iterating over vector<string>
As discussed in:
https://bitbucket.org/wlav/cppyy/issues/176

Cppyy has a new optimisation when iterating over a vector of
strings. The items returned by the iteration are now STL strings
and not Python strings, and their lifetime is tied to that of
the vector. However, lifelines are set for the strings if
the vector to iterate over has a reference count of 1, which is
what is achieved with the change in this commit.