[Exp PyROOT] Dynamically import the correct version of libROOTPythonizationsX_Y
In ROOT/__init__.py we use sys.version_info to get the version number of
the Python interpreter we're running with, and we append the unmbers to
'libROOTPythonizations' in order to find the correct shared library.

In PyROOTModule we need to define an init function with the same name of
the shared libraries, so in this case we use the preprocessor directives
to construct a function with the correct name.