[thread] Provide TThreadedObject() taking nslots:
It is ambiguous whether TThreadedObject should rely on IMT, or whether it is used by user scheduling.
The default of 64 threads was good enough for user scheduling - unless the user needs more threads.
The default of 64 threads is also good enough for IMT, when IMT was not yet enabled - unless IMT ends up using more slots.
Thus, the silent default value is often misleading, and setting it through a static before construction is not exactly cplusplusy.

Instead, plovide a constructor tag that passed the number of slots - or "IMT" if that is what is wanted.
Add a warning if IMT is assumed but IMT is off, rather than failing, if IMT later needs more than fgMaxSlots slots.