[Exp PyROOT][ROOT-10681] Provide both AddressOf and addressof
They both can be used to obtain the address of the address of
a C++ object from its Python proxy. In the case of addressof, the
address is returned as a value, while for AddressOf it is returned
as an indexable buffer.

This commits proposes to implement AddressOf in terms of addressof
(which is provided by cppyy), just by creating an array of one
position and storing the address in that position.