[DF] Store custom column type in RCustomColumnBase::fType
Before this patch, we lazily declared a type alias to the interpreter
for each custom column defined. Type aliases had the form
`<colName><ID>_type`, where ID was a globally unique identifier for
the custom column. Using such aliases in lambda signatures made it
impossible to re-use jitted lambdas across computation graphs because
even if logic was exactly the same, the type aliases would have different
IDs.

With this patch, we store the actual type of Define'd columns as a string
in RCustomColumnBase, which removes the need for the lazy type alias
declarations, simplifies RDF internal logic, and most importantly makes
sure that we only use "true"/resolved type names in lambda signatures,
which makes it possible to re-use jitted lambdas that take Define'd
columns as input across computation graphs.