[Exp PyROOT] Implement AddressOf using bytearray
The previous implementation of AddressOf introduced a requirement
of NumPy if Python<3.3, since array.array can't hold long longs
before that version. An array.array of longs would suffice in
Unix-based systems, where a long is 8 bytes, but not on Windows
64-bit, where a long is 4 bytes.

In this new implementation, we put the result of addressof in
a bytearray containing the bytes of a long long, no matter the
Python version. What this implies:
- The user can't rely anymore of AddressOf[0] to get the address
of the object, but that is precisely what addressof returns.
- The result of AddressOf can't be passed to a function that
expects a long long*, but it can still convert to void*.