[Math] Fix crash in TFoam.
[ROOT-10277] When creating a lot of datasets in RooFit, TFoam eventually
crashed, since it was using a TRefArray, which does some global UID
magic that overflows. Replacing the TRefArray by a simple vector of
indices fixes that.
This is possible, since the TRefArray only points to active elements in
another collection, so simply storing the index in that collection is
sufficient.

A ROOT-6.20 TFoam was created on lxplus, and written to a file. Reading
this file, and generating the 10 next random numbers is tested now.