TProcessID::RecursiveRemove: prevent spurrious error msg.
Prior to this change, TProcessID::RecursiveRemove would call TExMap::Remove without checking whether
the object was in that map or not.  In addition TExMap::Remove complains if it can not be found.
This resulted in spurious error message when the number of TProcessID increased passed 255 but some
of the object with lower index TProcessID were deleted.

See https://root-forum.cern.ch/t/resetting-tprocessid-objectcount-in-multihtreading-environment/38899
for a concrete example.