Load dependent libraries as part of AutoLoading.
Since the proper creation of the TClass will require the dictionary for all its component let
load all the dependent library too.  Note that even with explicit linking, the dictionary may
not be loaded as part of the 'main' library when the dictionary is a library separate from the
classes' code (see ATLAS setup for example)

This solves ROOT-10663.