[RF] Add capability to read ULong64_t + more into RooDataSets.
[ROOT-6882] RooFit wasn't able to attach to tree branches of various
types, but this was requested.
The buffers for various types were removed from the member list of
RooAbsReal, and were instead moved into a templated struct. This struct
is instantiated as needed, and takes care of buffering values coming
from trees and converting these to double.
This makes the list of members shorter, and converting things that
have been read from the tree doesn't require comparing strings, anymore.

Further, extending to other types is now trivial.