Allow to set the name in reproducible TFile
Rationale
---------

We create a number of small root files to serve as conditions data. The
algorithm tells the framework to save the data to be valid for a given interval of
time.

To de-duplicate files we use the md5 checksum: same checksum, same
content, so no new file needed. Technically we create the file as a
temporary file, calculate the checksum and use a directory structure
similar to git objects: if no file with the checksum as name exists,
rename, otherwise just use the existing name.

For this to work we need the md5 checksum independent of the name of the
TFile. And since we rename the file anyway it really doesn't matter what
TFile thinks its name is.

Implementation
--------------

This tries to be the minimal change to the current feature of
reproducible files (thanks again): We piggyback on the same flag but if
the flag also has a value instead of jut being present we use it as the
name to put in the file.