Fix potential GUI layout issues on Windows (#6077)
In some cases (e.g. in the parameters dialog of the fit panel) the size of the top level window (the dialog box - or TGTransientFrame) is wrongly calculated by the layout manager, and several widgets are invisible, forcing the user to manually resize the window in order to see and access all the controls inside the dialog. This is a side effect of the changes made a while back to support multiples displays on Windows. Since some part of the code should still be tested on multiple display, it is temporarily kept inside `#if 0` exclusion comment, but will be removed once successfully tested and validated.