Shift to standard C++ filestreams when generating _rdict files
On windows, there are instances when the rdict file is still in use when we try to remove it.

This happens due to some problem with file handling in LLVM5 which got resolved in D63453(https://reviews.llvm.org/D63453). But it was difficult to backport.

With this commit, we switch to the standard C++ library functions for generating rdict files and avoid using this particular API of LLVM.