[DF] Simplify reading of RDataSource columns
Before this commit, RDataSource columns were treated like a special
kind of Defined columns: they were registered in RBookedCustomColumn
and their contents were accessed via RCustomColumn::Get.

This commit removes the logic that was Define'ing ad-hoc columns
corresponding to the RDS columns. Instead, we store the RDS column value
pointers in a dedicated std::map and teach RDSColumnReader to directly
use that. Logic is simpler, we avoid an extra function call and an extra
copy upon data-source value accesses, and we move closer to implementing
column readers specialized for a given RDS implementation.