[cef] set different file paths in CefSettings
Different CEF files copied to lib/ or bin/ or Frameworks/ subdirectories
depending on the platform. But later CEF tries to access them from other
locations. To avoid duplication of such files, try to configure paths in
CefSettings.

Also configure paths to cef_main executable, which is started as
browser sub-process. Before path to executable was replaced in
SimpleApp::OnBeforeChildProcessLaunch handler