[DS] Add RDataSource::GetColumnReaders v2.0
The new version of the method allows the implementation of per-datasource,
lazy column readers that know how to communicate with the concrete
associated datasource type.

As a consequence, datasource implementations are not required to load
values of all required columns when RDataSource::SetEntry is called, but
the loading can be delayed to the moment in which a value is actually
needed. In case of strict upstream Filters, this should result in
less work performed and therefore improved runtimes.