[RF] Don't copy expensive object cache pointers in RooAbsArg.
If an object inside a workspace is cloned, its pointer to the expensive
object cache inside the original workspace is copied. This creates a
dangling pointer when the original workspace is deleted. This excludes
the pointer from cloning, so clones will store their objects in RooFit's
global cache or (after they have been assigned to a new workspace) in
their new workspace.

Fixes root-project/root#6529