[DF] Jit and exec Cache together with the rest
Before this commit, a Cache call without template parameters would
trigger jitting and execution of the corresponding typed Cache call
on the spot, separately from the rest of RDF's jitting.

With these changes the code needed by Cache is merged with the rest of
the code booked for jitting. Jitting is then triggered on the spot,
because a jitted Cache call needs the jitted code to produce its return
value.

This resolves the Cache-related part of ROOT-9790, "Lazy jitting of
Cache and Snapshot".