[RF] Implement retrieval of batches in RooDataSet, RooDataHist.
To efficiently use the new batch interface, data stores should be able
to return RunContexts where the spans are already mapped to the objects
that own the data.
The RooVectorDataStore can now generate those natively. It maps the data
columns to both the instances in the dataset as well as to instances
whose buffers were attached to the dataset. This means that no mapping
between "variable x in dataset" to "variable x of computation graph" has
to be made during evaluation.