[PyROOT][6552] Import libcppyy during initialization of libROOTPythonizations
When doing `import ROOT`, libcppyy is imported before libROOTPythonizations,
so the change proposed by this commit is not necessary.

However, as reported in #6552, if ROOT objects are being deserialized,
it is possible that libROOTPythonizations is initialized before
libcppyy has been imported as a module, which causes a crash. This
commits ensures libcppyy is imported before calling into it during
the initialization of libROOTPythonizations.