Add possibility to get ROC curve for training dataset
- extend API in Factory to get the ROC curve for the training data
- fimprove ResultsClassification to fill at the same time both output score
results and classID information. Before the classID info was set only later in
MethodBase::EvaluateAllMethod and done only for the testing data set.
Now it is done when the ResultsClassification object is filled
(in MethodBase::AddClassifierOutput).
With these changes one can get a complete results for both training and testing data and being able to retrieve ROC curve in both cases

- fix also the code formatting for Factory class. Indentation was not correct