[RF] Delete copy constructor of RunContext.
The RunContext owns memory, and spans are used to pass data around. If a
RunContext would be copied, all spans would have to be relocated, as
they would all contain dangling pointers.
Until a real need for copies arises, only moves are allowed.

This fixes a failure where an older version of the STL preferred a copy
over a move.