Avoid a duplicate state in Funciton Minimum
When using stratgey 0 (i.e. when Hesse is not computed at the end of the minimization), the same state is added 2 times in the FunctionMinimum class.
This is now fixed by adding a flag runHessian in the VariableMetric code.

Add another improvement: do not compute Hessian during the iteration.
ALso when not using debug mode the Hessian, which requires a matrix inversion, was computed anyway for each iteration. This is now fixed