[metacling] Also fwd decl templt spec args:
rootmap needs { decls } entries such as
```
template <> class less<edm::AJet>;
```
to be able to load the right library containing a given specialization.
But for this to be parsable, the template arguments of that specialization
must also be forward declared. Re-use existing template-arg-fwd-decl code,
and hook it into FwdDeclIfTmplSpec.

Fixes roottest/cling/dict/fwd-decl-stdless/lessyDict.rootmap in the PCH
case, and likely CMS's issue of a similar kind - TBC.