[RF] Check if ranges are not overlapping in multi-range fit
Events are double counted if one accidentally defines overlapping ranges
and uses them in a multi-range fit. This happened for example in Jira
issue ROOT-9548 where the whole dataset was double counted, leading to
the parameter uncertainties being underestimated by a factor of sqrt(2).

That situation should be avoided. This commit introduces a check for
overlapping ranges before the multi-range likelihood is created.